# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/posts": {
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created"
          }
        },
        "summary": "Create a post (content calendar entry)",
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "PostsController_findAll",
        "parameters": [
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "draft",
                "review",
                "approved",
                "scheduled",
                "published",
                "analyzing",
                "underperforming",
                "completed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List posts (optionally by campaign/status) — content calendar feed",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{id}": {
      "get": {
        "operationId": "PostsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a post with metrics, analysis, and recommendations",
        "tags": [
          "posts"
        ]
      },
      "patch": {
        "operationId": "PostsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a post",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/api/posts/{id}/publish": {
      "post": {
        "operationId": "PostsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Published post"
          }
        },
        "summary": "Publish a post (simulates going live). The autonomous metric-checker then pulls mock analytics on a cron and runs the agent loops automatically.",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/api/recommendations": {
      "get": {
        "operationId": "RecommendationsController_list",
        "parameters": [
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "postId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "open",
                "accepted",
                "dismissed",
                "superseded"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List agent recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/recommendations/{id}/status": {
      "patch": {
        "operationId": "RecommendationsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Accept / dismiss a recommendation",
        "tags": [
          "recommendations"
        ]
      }
    },
...
    "/api/integrations/{provider}/connect": {
      "post": {
        "operationId": "IntegrationsController_connect",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Connect a content-calendar app",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/integrations/{provider}/disconnect": {
      "post": {
        "operationId": "IntegrationsController_disconnect",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Disconnect a content-calendar app",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/integrations/{provider}/sync": {
      "post": {
        "operationId": "IntegrationsController_sync",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Push the calendar out + pull external events",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/integrations/{provider}/publish": {
      "post": {
        "operationId": "IntegrationsController_publish",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Publish a post through a connected publishing account (gated on credentials)",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/missions": {
      "post": {
        "operationId": "MissionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mission created"
          }
        },
        "summary": "Create a mission (the unit of goal-driven work)",
        "tags": [
          "missions"
        ]
      },
      "get": {
        "operationId": "MissionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List missions",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{id}": {
      "get": {
        "operationId": "MissionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a mission with campaigns + agent runs",
        "tags": [
          "missions"
        ]
      },
      "patch": {
        "operationId": "MissionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{id}/progress": {
      "get": {
        "operationId": "MissionsController_progress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mission goal progress vs target metric",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{id}/run": {
      "post": {
        "operationId": "MissionsController_run",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Run the mission: Planner → Strategy (creates campaign) → Content → Compliance → Swarm → Approval gates. Returns the mission with its full agent-run timeline.",
        "tags": [
          "missions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hiive Agentic Marketing OS",
    "description": "Mission-driven, multi-agent marketing operating system: RAG memory, campaign intelligence, swarm simulation, autonomous monitoring, human-in-the-loop approvals, and a continuous learning flywheel.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "example": "campaign id"
          },
          "platform": {
            "type": "string",
            "example": "LinkedIn"
          },
          "copy": {
            "type": "string",
            "example": "Thinking about liquidity before your IPO? ..."
          },
          "mediaUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "review",
              "approved",
              "scheduled",
              "published",
              "analyzing",
              "underperforming",
              "completed"
            ]
          },
          "scheduledAt": {
            "type": "string",
            "example": "2026-06-10T09:00:00.000Z"
          }
        },
        "required": [
          "campaignId",
          "platform",
          "copy"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "example": "campaign id"
          },
          "platform": {
            "type": "string",
            "example": "LinkedIn"
          },
          "copy": {
            "type": "string",
            "example": "Thinking about liquidity before your IPO? ..."
          },
          "mediaUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "review",
              "approved",
              "scheduled",
              "published",
              "analyzing",
              "underperforming",
              "completed"
            ]
          },
          "scheduledAt": {
            "type": "string",
            "example": "2026-06-10T09:00:00.000Z"
          }
        }
      },
...
      "CreateMissionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Increase sell-side founder leads by 30%"
          },
          "objective": {
            "type": "string",
            "example": "Drive more inbound founders who want pre-IPO liquidity"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "targetMetric": {
            "type": "object",
            "example": {
              "metric": "founder_leads",
              "baseline": 100,
              "target": 130,
              "unit": "leads"
            }
          },
          "audience": {
            "type": "string",
            "example": "Startup founders and early employees"
          }
        },
        "required": [
          "title",
          "objective",
          "targetMetric"
        ]
      },
      "UpdateMissionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Increase sell-side founder leads by 30%"
          },
          "objective": {
            "type": "string",
            "example": "Drive more inbound founders who want pre-IPO liquidity"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "targetMetric": {
            "type": "object",
            "example": {
              "metric": "founder_leads",
              "baseline": 100,
              "target": 130,
              "unit": "leads"
            }
          },
          "audience": {
            "type": "string",
            "example": "Startup founders and early employees"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "planning",
              "executing",
              "evaluating",
              "reflecting",
              "awaiting_approval",
              "completed",
              "failed"
            ]
          }
        }
      },
...
    }
  }
}